backend/requirements.txt
google-genai
python-dotenv
opencv-python-headless
nltk
matplotlib
uvicorn
fastapi
python-multipart